git-svn-id: svn://svn.code.sf.net/p/ponykart/code/trunk@447 ba97e771-f6de-4074-9afc-763755853dd4

+++ Ponykart/Physics/Collision Reports/CollisionReporter.cs
@@ -40,7 +40,7 @@ namespace Ponykart.Physics {
 		/// <summary>
 		/// remember to update this!
 		/// </summary>
-		static readonly int HIGHEST_BIT_IN_COLLISION_GROUPS = 64;
+		static readonly byte HIGHEST_BIT_IN_COLLISION_GROUPS = 64;
 
 		#endregion -------------------- HEY THIS IS IMPORTANT ---------------------
